+++ src/task/CleanStoppedSessionTask.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import AudioVideoControllerState from '../audiovideocontroller/AudioVideoControllerState';
@@ -55,11 +55,6 @@ export default class CleanStoppedSessionTask extends BaseTask {
       this.context.turnCredentials = null;
       this.context.videoSubscriptions = null;
       this.context.transceiverController.reset();
-      this.context.mediaStreamBroker.releaseMediaStream(this.context.activeAudioInput);
-      this.context.activeAudioInput = null;
-      this.context.mediaStreamBroker.releaseMediaStream(this.context.activeVideoInput);
-      this.context.activeVideoInput = null;
-      this.context.realtimeController.realtimeSetLocalAudioInput(null);
 
       const tile = this.context.videoTileController.getLocalVideoTile();
       if (tile) {